% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_covid_annotation.R
\name{jp_covid_annotation}
\alias{jp_covid_annotation}
\title{Add COVID-19 pandemic annotation to Datawrapper charts}
\usage{
jp_covid_annotation(
  chart_id,
  date = "2020/03",
  covid_text_y = NA,
  text_color = "#888",
  text_bold = FALSE,
  line_color = "#888",
  line_type = "solid",
  line_width = 1
)
}
\arguments{
\item{chart_id}{Datawrapper chart id.}

\item{date}{Position of the line and text annotations on the x-axis.
Defaults to "2020/03", in which Indonesia reported its
first two confirmed coronavirus cases.}

\item{covid_text_y}{Position of the COVID-19 text annotation
on the y-axis.}

\item{text_color}{Color of the text annotation. Defaults to "#888".}

\item{text_bold}{Logical. Defaults to \code{FALSE}. When \code{TRUE},
it makes the text annotation bold.}

\item{line_color}{Color of the line annotation. Defaults to "#888".}

\item{line_type}{Line type for the line annotation.
Defaults to "solid". "dotted" gives dotted line. "dashed" gives dashed line.}

\item{line_width}{Integer. Width of the line annotation, ranging from 1 to 3.
Defaults to 1.}
}
\description{
Add line and text annotations for COVID-19 pandemic to Datawrapper charts.
}
\examples{
\dontrun{
jp_covid_annotation(
  chart_id = "4BcD3",
  covid_text_y = 6,
  text_color = "#ee493a",
  text_bold = TRUE,
  line_color = "#ee493a",
  line_type = "dashed",
  line_width = 2
)
}

}
\seealso{
\code{\link[DatawRappr]{dw_edit_chart}} for the underlying function.
}
