% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_dw_c19_annotation.R
\name{jp_dw_c19_annotation}
\alias{jp_dw_c19_annotation}
\title{Add COVID-19 annotations}
\usage{
jp_dw_c19_annotation(
  chart_id,
  x_position = "2020/03",
  y_position = NA,
  text_bold = FALSE,
  lty = "dotted",
  lwd = 2
)
}
\arguments{
\item{chart_id}{String. Datawrapper chart id.}

\item{x_position}{String. The line and text annotations’ position on the x-axis.
Defaults to “2020/03”.}

\item{y_position}{Integer. The text annotation’s position on the y-axis.}

\item{text_bold}{Logical. Defaults to FALSE.}

\item{lty}{String. The type of line annotation. Defaults to “dotted”.
Also takes “solid” and “dashed”.}

\item{lwd}{Integer. The width of line annotation, which ranges from 1 to 3.
Defaults to 2.}
}
\description{
Add line and text annotations to highlight the COVID-19 pandemic in
a chart that maps a date variable on the x-axis and a continuous variable
on the y-axis.
}
\examples{
\dontrun{
jp_dw_c19_annotation(
  chart_id = "4BcD3",
  y_position = 6,
  text_bold = TRUE,
  lty = "dashed",
  lwd = 2
)
}

}
\seealso{
\code{\link[DatawRappr]{dw_edit_chart}} for the underlying function.
}
