% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_dw_scale_x.R
\name{jp_dw_scale_x}
\alias{jp_dw_scale_x}
\title{Modify x-axis}
\usage{
jp_dw_scale_x(chart_id, format = "auto", axis_line = FALSE)
}
\arguments{
\item{chart_id}{String. Datawrapper chart id.}

\item{format}{String. Defaults to “auto”. Also takes “monthly” and
“quarterly” for date variables.}

\item{axis_line}{Logical. Defaults to FALSE. When TRUE, adds a black axis line.}
}
\description{
Customize the axis tick format and add an axis line to a chart with a y-axis
that doesn’t start at zero.
}
\examples{
\dontrun{
jp_dw_scale_x(
  chart_id = "4BcD3",
  format = "quarterly",
  axis_line = TRUE
)
}

}
\seealso{
\code{\link[DatawRappr]{dw_edit_chart}} for the underlying function.
}
