% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_dw_theme.R
\name{jp_dw_theme}
\alias{jp_dw_theme}
\title{Apply The Jakarta Post theme}
\usage{
jp_dw_theme(chart_id, author, footnote = "", logo = TRUE, ...)
}
\arguments{
\item{chart_id}{String. Datawrapper chart id.}

\item{author}{String. Who creates the chart?}

\item{footnote}{String. Footnote to the chart.}

\item{logo}{Logical. Defaults to TRUE. When FALSE, the function does not add
The Jakarta Post’s logo to the chart.}

\item{...}{Other arguments passed on to \code{\link[DatawRappr]{dw_edit_chart}},
including \code{intro}, \code{source_name} and \code{source_url}.}
}
\description{
Add a prefix to the byline, the Post’s logo to the chart and modify axes and
the grid lines, in addition to other minor changes.
}
\examples{
\dontrun{
jp_dw_theme(
  chart_id = "4BcD3",
  author = "Leonard Hofstadter",
  intro = "Indonesia's GDP per capita (in US dollars)",
  source_name = "Gapminder"
)
}

}
\seealso{
\code{\link[DatawRappr]{dw_edit_chart}} for the underlying function.
}
