% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_scale_y.R
\name{jp_scale_y}
\alias{jp_scale_y}
\title{Customize y-axis of Datawrapper charts}
\usage{
jp_scale_y(
  chart_id,
  y,
  scale_y_max = max(y),
  scale_y_min_rule = "min",
  scale_y_round,
  scale_y_increment
)
}
\arguments{
\item{chart_id}{Datawrapper chart id.}

\item{y}{String. Name of variable mapped to the y-axis.}

\item{scale_y_max}{Integer. The maximum value of the y-axis range.
Defaults to the highest value in the variable.}

\item{scale_y_min_rule}{Minimum value of the y-axis range,
either as an integer or a string. Defaults to "min", which takes
the lowest value in the variable. "truncated" leaves some
one-third space of the plot area empty between the lowest value
in the variable and the start of the scale.}

\item{scale_y_round}{Number of decimal places for all values in the y-axis.
Negative values round to a power of ten.}

\item{scale_y_increment}{Increment of the y-axis breaks.}
}
\description{
Customize the y-axis limits and major breaks of Datawrapper charts.
}
\examples{
\dontrun{
jp_scale_y(
  chart_id = "4BcD3",
  y = "lifeExp",
  scale_y_min_rule = 0,
  scale_y_round = -2, # Rounds to the nearest hundred
  scale_y_increment = 20
)
}

}
\seealso{
\code{\link[DatawRappr]{dw_edit_chart}} and \code{\link[base]{round}}
for the underlying functions.
}
