% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jakpost_style.R
\name{jakpost_style}
\alias{jakpost_style}
\title{Apply The Jakarta Post theme to Datawrapper charts}
\usage{
jakpost_style(chart_id, author, footnote = "", logo = TRUE, ...)
}
\arguments{
\item{chart_id}{Datawrapper chart id.}

\item{author}{Byline of the chart.}

\item{footnote}{Footnote to the chart.}

\item{logo}{Logical. Defaults to \code{TRUE}.
When \code{FALSE}, it removes The Jakarta Post's logo from the chart.}

\item{...}{Other arguments passed on to \code{\link[DatawRappr]{dw_edit_chart}},
including \code{intro}, \code{source_name} and \code{source_url} for chart subtitle,
source and source's link, respectively.}
}
\description{
Use the The Jakarta Post custom theme to edit Datawrapper chart properties.
}
\examples{
\dontrun{
jakpost_style(
  chart_id = "4BcD3",
  author = "Dzulfiqar Fathur Rahman",
  intro = "Indonesia's gross domestic product (GDP) per capita (in US dollars)",
  source_name = "Gapminder"
)
}

}
\seealso{
\code{\link[DatawRappr]{dw_edit_chart}} for the underlying function.
}
